auth & status added

+++ controllers/feed.js
@@ -3,6 +3,7 @@ const path = require('path');
 
 const { validationResult } = require('express-validator');
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
     const currentPage = req.query.page || 1;
@@ -48,19 +49,27 @@ exports.createPost = (req, res, next) => {
     const imageUrl = req.file.path;
     const title = req.body.title;
     const content = req.body.content;
-    // Create post in db
+    let creator;
     const post = new Post({
         title: title,
         content: content,
         imageUrl: imageUrl,
-        creator: { name: 'Alex' },
+        creator: req.userId, // I saved an user id in is-auth.js / req.userId = decodedToken.userId;
     });
     post.save()
         .then((result) => {
-            console.log(result);
+            return User.findById(req.userId);
+        })
+        .then((user) => {
+            creator = user; // create this variable to send a creator info to the response.
+            user.posts.push(post); // adding new post to the user in MongoDB
+            return user.save();
+        })
+        .then((result) => {
             res.status(201).json({
                 message: 'Post created successfully!',
-                post: result,
+                post: post,
+                creator: { id: creator._id, name: creator.name },
             });
         })
         .catch((err) => {
@@ -118,6 +127,11 @@ exports.updatePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized!');
+                error.statusCode = 403;
+                throw error;
+            }
             if (imageUrl !== post.imageUrl) {
                 clearImage(post.imageUrl);
             }
@@ -146,12 +160,23 @@ exports.deletePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized!');
+                error.statusCode = 403;
+                throw error;
+            }
             // Check logged user
             clearImage(post.imageUrl);
             return Post.findByIdAndRemove(postId);
         })
         .then((result) => {
-            console.log(result);
+            return User.findById(req.userId);
+        })
+        .then((user) => {
+            user.posts.pull(postId);
+            return user.save();
+        })
+        .then((result) => {
             res.status(200).json({ message: 'Deleted post.' });
         })
         .catch((err) => {
